Optimize simultaneity calculation
Careful: Previously, the original profile was not changed,
only all the copies of a profile.
Now the first profile is shifted, too, if a standard
deviation sigma is defined.
The number 'copies' in the yaml config does not define the
additional copies anymore, but the total number of resulting
columns. For the same output as before, it needs to be
increased by 1.