Restrict cx_freeze to 6.14.* in workflow
While the workflows run fine, the resulting executables
do not work due to dll link errors. These are related to
mkl and numpy most of the time, and how cx_freeze packages
those dependencies.
Might have to remove the restriction in the future, when
other versions of e.g. numpy are used.