Accept transposed Excel houses table as input
Houses can either be defined as a dictionary in the
configuration.yaml, and/or the yaml may point to an Excel file
containing the table of houses. A new setting 'transpose' has
been added, which just transposes the DataFrame directly after
loading the Excel file. It can be used with 'True', if the
input table contains the buildings as rows (instead of columns).

houses_table_file:
    file: ./VDI_4655_houses_example.xlsx
    transpose: True  # Flip rows and columns of the input table
    kwargs:  # optional keyword arguments for pandas read_excel()
        sheet_name: Example