Implement BDEW profiles from demandlib as new source
When the settings option 'use_demandlib' is set to True, commercial
profiles are now calculated with the BDEW implementation from
demandlib. Since different building type notations are used for
thermal and electrical profiles, the expected notation of the
'house_type' key has changed to the format 'GHD/G1', denoting the
respective types to use for the profiles.

For a list of the heat and electricity types, please refer to
the demandlib documentation
https://demandlib.readthedocs.io/en/latest/bdew.html

For single- and multi-family homes the VDI 4655 implementation in
demandlib is used, which is faster than the lpagg implementation.